don't reuse error types across stacks
