+++ linkerd/app/outbound/src/http/concrete.rs
@@ -1,11 +1,23 @@
 use super::{Concrete, Endpoint};
-use crate::{endpoint, logical::ConcreteError, stack_labels, Outbound};
+use crate::{endpoint, stack_labels, Outbound};
 use linkerd_app_core::{
-    proxy::{api_resolve::Metadata, core::Resolve, http},
+    proxy::{
+        api_resolve::{ConcreteAddr, Metadata},
+        core::Resolve,
+        http,
+    },
     svc, Error,
 };
 use std::time;
 use tracing::info_span;
+
+#[derive(Debug, thiserror::Error)]
+#[error("concrete service {addr}: {source}")]
+pub struct ConcreteError {
+    addr: ConcreteAddr,
+    #[source]
+    source: Error,
+}
 
 impl<N> Outbound<N> {
     /// Builds a [`svc::NewService`] stack that builds buffered HTTP load
@@ -72,11 +84,24 @@ impl<N> Outbound<N> {
     }
 }
 
+// === impl Concrete ===
+
 impl svc::Param<http::balance::EwmaConfig> for Concrete {
     fn param(&self) -> http::balance::EwmaConfig {
         http::balance::EwmaConfig {
             default_rtt: time::Duration::from_millis(30),
             decay: time::Duration::from_secs(10),
+        }
+    }
+}
+
+// === impl ConcreteError ===
+
+impl<T: svc::Param<ConcreteAddr>> From<(&T, Error)> for ConcreteError {
+    fn from((target, source): (&T, Error)) -> Self {
+        Self {
+            addr: target.param(),
+            source,
         }
     }
 }
